web: skip backend requests if feature flag override exists
