+++ client/web/src/featureFlags/lib/feature-flag-local-overrides.ts
@@ -8,3 +8,21 @@ export const setFeatureFlagOverride = (flagName: string, value: boolean): void =
 
 export const removeFeatureFlagOverride = (flagName: string): void =>
     localStorage.removeItem(buildFlagOverrideKey(flagName))
+
+export const getFeatureFlagOverrideValue = (flagName: string): boolean | null => {
+    const overriddenValue = getFeatureFlagOverride(flagName)
+
+    if (overriddenValue === null) {
+        return null
+    }
+
+    if (['true', 1].includes(overriddenValue)) {
+        return true
+    }
+
+    if (['false', 0].includes(overriddenValue)) {
+        return false
+    }
+
+    throw new Error(`FeatureFlagOverride "${flagName}" has invalid value: ${overriddenValue}`)
+}
